chore(compilers): remove Python Dockerfile and update compile.py to add command line argument support
The Python Dockerfile was removed as it is no longer needed. The compile.py script was updated to add support for command line arguments using the argparse module. The script now accepts an input Uniword code as a positional argument and an optional output file path as a command line option. If an output file path is provided, the generated Python code will be written to that file. Otherwise, the generated Python code will be printed to the console.